+++ src/pixelator/graph/backends/implementations/_networkx.py
@@ -733,7 +733,9 @@ def pmds_layout(
 
     The algorithm is similar to classical multidimensional scaling (MDS), but uses only
     a smalls set of random pivot nodes. The algorithm is considerably faster than MDS
-    and therefore scales better to large graphs.
+    and therefore scales better to large graphs. The topology of resulting layouts are
+    deterministic for a given seed, but may be mirrored across different systems due to
+    variations in floating-point precision.
 
     .. [1] Brandes U, Pich C. Eigensolver Methods for Progressive Multidimensional
         Scaling of Large Data. International Symposium on Graph Drawing, 2007.
@@ -795,7 +797,7 @@ def pmds_layout(
 
     # Compute SVD and use distances to compute coordinates for all nodes
     # in an abstract cartesian space
-    _, _, Vh = sp.sparse.linalg.svds(D_pivs_centered, k=dim)
+    u, _, Vh = sp.sparse.linalg.svds(D_pivs_centered, k=dim, random_state=seed)
     coordinates = D_pivs_centered @ np.transpose(Vh)
 
     coordinates = {node_list[i]: coordinates[i, :] for i in range(coordinates.shape[0])}